+++ tests/integration/helpers/core.ts
@@ -1,6 +1,7 @@
 import path from 'node:path';
 import {MessageChannel} from 'node:worker_threads';
 import {FastifyController, MapeoManager} from '@comapeo/core';
+import type {MapeoProjectApi} from '@comapeo/ipc';
 import {
   closeMapeoClient,
   createMapeoClient,
@@ -8,7 +9,11 @@ import {
 } from '@comapeo/ipc';
 import {KeyManager} from '@mapeo/crypto';
 import Fastify from 'fastify';
+import {randomBytes} from 'node:crypto';
+import {pEvent} from 'p-event';
+import pEvery from 'p-every';
 import RAM from 'random-access-memory';
+import {MEMBER_ROLE_ID} from '../../../src/frontend/sharedTypes';
 
 const COMAPEO_CORE_PKG_FOLDER = path.dirname(
   require.resolve('@comapeo/core/package.json'),
@@ -70,4 +75,92 @@ export function setUpIPC({manager}: {manager: MapeoManager}) {
       port2.close();
     },
   };
+}
+
+export async function connectPeers(
+  managers: ReadonlyArray<MapeoManager>,
+): Promise<() => Promise<void>> {
+  await tellPeersAboutEachOther(managers);
+  await waitForPeersToBeConnected(managers);
+  return () => stopPeerDiscovery(managers);
+}
+
+async function tellPeersAboutEachOther(
+  managers: ReadonlyArray<MapeoManager>,
+): Promise<void> {
+  await Promise.all(
+    managers.map(async manager => {
+      const {name, port} = await manager.startLocalPeerDiscoveryServer();
+      for (const otherManager of managers) {
+        if (otherManager === manager) continue;
+        otherManager.connectLocalPeer({address: '127.0.0.1', name, port});
+      }
+    }),
+  );
+}
+
+async function waitForPeersToBeConnected(
+  managers: ReadonlyArray<MapeoManager>,
+): Promise<void> {
+  const deviceIds = new Set(managers.map(m => m.deviceId));
+
+  const isDone = async (): Promise<boolean> =>
+    pEvery(managers, async manager => {
+      const unconnectedDeviceIds = new Set(deviceIds);
+      unconnectedDeviceIds.delete(manager.deviceId);
+      for (const peer of await manager.listLocalPeers()) {
+        if (peer.status === 'connected') {
+          unconnectedDeviceIds.delete(peer.deviceId);
+        }
+      }
+      return unconnectedDeviceIds.size === 0;
+    });
+
+  if (await isDone()) return;
+
+  return new Promise(res => {
+    const onLocalPeers = async () => {
+      if (await isDone()) {
+        for (const manager of managers) {
+          manager.off('local-peers', onLocalPeers);
+        }
+        res();
+      }
+    };
+    for (const manager of managers) manager.on('local-peers', onLocalPeers);
+  });
+}
+
+async function stopPeerDiscovery(
+  managers: ReadonlyArray<MapeoManager>,
+): Promise<void> {
+  await Promise.all(
+    managers.map(manager =>
+      manager.stopLocalPeerDiscoveryServer({force: true}),
+    ),
+  );
+}
+
+export async function inviteToProject(
+  project: MapeoProjectApi,
+  invitee: MapeoManager,
+): Promise<void> {
+  const inviteId = randomBytes(32);
+
+  const inviteeInvitePromise = pEvent(
+    invitee.invite,
+    'invite-received',
+    invite => Buffer.from(invite.inviteId, 'hex').equals(inviteId),
+  );
+
+  await Promise.all([
+    project.$member.invite(invitee.deviceId, {
+      roleId: MEMBER_ROLE_ID,
+      __testOnlyInviteId: inviteId,
+    }),
+    (async () => {
+      const invite = await inviteeInvitePromise;
+      await invitee.invite.accept(invite);
+    })(),
+  ]);
 }
